Dawini2025/requirements.txt
<<<<<<< HEAD
Django>=4.2,<5.0
djangorestframework<3.15
djangorestframework-simplejwt<5.3
django-cors-headers
mongoengine
pymongo
joblib
scikit-learn
pandas
numpy
channels>=4.0
channels-redis
python-dotenv
Pillow
bcrypt
=======
Django==5.2.8
djangorestframework==3.16.1
pymongo==4.15.4
djangorestframework-simplejwt==5.5.1
firebase-admin==7.1.0
pyfcm==2.1.0
celery==5.5.3
redis==7.0.1
drf-spectacular==0.29.0
scikit-learn==1.7.2
pandas==2.3.3
numpy==2.3.4
python-dotenv==1.0.1
Pillow==11.1.0
django-cors-headers==4.6.0
gunicorn==23.0.0
whitenoise==6.9.0
>>>>>>> 6d6560c2f8c56409972f4c856a7d4e04e3ee9e7a
